pyproject.toml
[project]
name = "media-trimmer"
version = "0.1.0"
description = ""
authors = [
    {name = "John Zhou",email = "johnzhou.engineer@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyQt5",
    "python-vlc",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
